feat(ER-2785): accounts, resources, roles and users data sources (#49)

+++ internal/client/entitle.gen.go
@@ -2487,10 +2487,10 @@ type AccessReviewForwardsIndexParams struct {
 
 // AccountsIndexParams defines parameters for AccountsIndex.
 type AccountsIndexParams struct {
-	Page          *float32 `form:"page,omitempty" json:"page,omitempty"`
-	PerPage       *float32 `form:"perPage,omitempty" json:"perPage,omitempty"`
-	IntegrationId string   `form:"integrationId" json:"integrationId"`
-	Search        *string  `form:"search,omitempty" json:"search,omitempty"`
+	Page          *int    `form:"page,omitempty" json:"page,omitempty"`
+	PerPage       *int    `form:"perPage,omitempty" json:"perPage,omitempty"`
+	IntegrationId string  `form:"integrationId" json:"integrationId"`
+	Search        *string `form:"search,omitempty" json:"search,omitempty"`
 }
 
 // AgentTokensIndexParams defines parameters for AgentTokensIndex.
@@ -2544,25 +2544,25 @@ type PoliciesIndexParams struct {
 
 // ResourcesIndexParams defines parameters for ResourcesIndex.
 type ResourcesIndexParams struct {
-	Page          *float32 `form:"page,omitempty" json:"page,omitempty"`
-	PerPage       *float32 `form:"perPage,omitempty" json:"perPage,omitempty"`
-	Search        *string  `form:"search,omitempty" json:"search,omitempty"`
-	IntegrationId string   `form:"integrationId" json:"integrationId"`
+	Page          *int    `form:"page,omitempty" json:"page,omitempty"`
+	PerPage       *int    `form:"perPage,omitempty" json:"perPage,omitempty"`
+	Search        *string `form:"search,omitempty" json:"search,omitempty"`
+	IntegrationId string  `form:"integrationId" json:"integrationId"`
 }
 
 // RolesIndexParams defines parameters for RolesIndex.
 type RolesIndexParams struct {
-	Page       *float32           `form:"page,omitempty" json:"page,omitempty"`
-	PerPage    *float32           `form:"perPage,omitempty" json:"perPage,omitempty"`
+	Page       *int               `form:"page,omitempty" json:"page,omitempty"`
+	PerPage    *int               `form:"perPage,omitempty" json:"perPage,omitempty"`
 	ResourceId openapi_types.UUID `form:"resourceId" json:"resourceId"`
 	Search     *string            `form:"search,omitempty" json:"search,omitempty"`
 }
 
 // UsersIndexParams defines parameters for UsersIndex.
 type UsersIndexParams struct {
-	Page    *float32 `form:"page,omitempty" json:"page,omitempty"`
-	PerPage *float32 `form:"perPage,omitempty" json:"perPage,omitempty"`
-	Search  *string  `form:"search,omitempty" json:"search,omitempty"`
+	Page    *int    `form:"page,omitempty" json:"page,omitempty"`
+	PerPage *int    `form:"perPage,omitempty" json:"perPage,omitempty"`
+	Search  *string `form:"search,omitempty" json:"search,omitempty"`
 }
 
 // UsersAccountsIndexParams defines parameters for UsersAccountsIndex.
